requirements.txt
fastapi
uvicorn
pandas
python-dotenv
copernicusmarine
numpy